Refactored actions to static form functions
